refactor: extract arena_logger into standalone header
- New file: basic/arena_logger.h — self-contained profiling module
- dbg.h delegates to arena_logger.h via single include
- arena.h untouched — sees logger declarations through
  arena.h → dbg.h → arena_logger.h include chain
- All 16 arena tests pass
